Moved capture/chooser logic to the images fragment
