+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductDetailFragment.kt
@@ -1,11 +1,7 @@
 package com.woocommerce.android.ui.products
 
-import android.Manifest.permission
-import android.app.Activity
-import android.app.Activity.RESULT_OK
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.Menu
@@ -30,7 +26,6 @@ import com.woocommerce.android.analytics.AnalyticsTracker.Stat.PRODUCT_DETAIL_IM
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat.PRODUCT_DETAIL_SHARE_BUTTON_TAPPED
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat.PRODUCT_DETAIL_VIEW_AFFILIATE_TAPPED
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat.PRODUCT_DETAIL_VIEW_EXTERNAL_TAPPED
-import com.woocommerce.android.media.MediaUploadUtils
 import com.woocommerce.android.model.Product
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.base.UIMessageResolver
@@ -38,9 +33,7 @@ import com.woocommerce.android.ui.products.ProductDetailViewModel.ProductWithPar
 import com.woocommerce.android.ui.products.ProductType.EXTERNAL
 import com.woocommerce.android.ui.products.ProductType.GROUPED
 import com.woocommerce.android.ui.products.ProductType.VARIABLE
-import com.woocommerce.android.util.FeatureFlag
 import com.woocommerce.android.util.StringUtils
-import com.woocommerce.android.util.WooPermissionUtils
 import com.woocommerce.android.widgets.SkeletonView
 import com.woocommerce.android.widgets.WCProductImageGalleryView.OnGalleryImageClickListener
 import dagger.android.support.AndroidSupportInjection
@@ -51,16 +44,6 @@ import org.wordpress.android.util.HtmlUtils
 import javax.inject.Inject
 
 class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
-    companion object {
-        private const val MENU_ID_CHOOSE_PHOTO = 2
-        private const val MENU_ID_CAPTURE_PHOTO = 3
-
-        private const val REQUEST_CODE_CHOOSE_PHOTO = Activity.RESULT_FIRST_USER
-        private const val REQUEST_CODE_CAPTURE_PHOTO = REQUEST_CODE_CHOOSE_PHOTO + 1
-
-        private const val KEY_CAPTURED_PHOTO_URI = "captured_photo_uri"
-    }
-
     private enum class DetailCard {
         Primary,
         PricingAndInventory,
@@ -77,14 +60,10 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
     private var isVariation = false
     private var imageHeight = 0
     private val skeletonView = SkeletonView()
-    private var capturedPhotoUri: Uri? = null
 
     private val navArgs: ProductDetailFragmentArgs by navArgs()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        savedInstanceState?.let { bundle ->
-            capturedPhotoUri = bundle.getParcelable(KEY_CAPTURED_PHOTO_URI)
-        }
         setHasOptionsMenu(true)
         return inflater.inflate(R.layout.fragment_product_detail, container, false)
     }
@@ -110,11 +89,6 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
         initializeViewModel()
     }
 
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        outState.putParcelable(KEY_CAPTURED_PHOTO_URI, capturedPhotoUri)
-    }
-
     private fun initializeViewModel() {
         viewModel = ViewModelProviders.of(this, viewModelFactory).get(ProductDetailViewModel::class.java).also {
             setupObservers(it)
@@ -136,20 +110,8 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
             shareProduct(it)
         })
 
-        viewModel.chooseProductImage.observe(this, Observer {
-            chooseProductImage()
-        })
-
-        viewModel.captureProductImage.observe(this, Observer {
-            captureProductImage()
-        })
-
         viewModel.showSnackbarMessage.observe(this, Observer {
             uiMessageResolver.showSnack(it)
-        })
-
-        viewModel.isUploadingProductImage.observe(this, Observer {
-            showUploadImageProgress(it)
         })
 
         viewModel.exit.observe(this, Observer {
@@ -170,10 +132,6 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
     override fun onCreateOptionsMenu(menu: Menu?, inflater: MenuInflater?) {
         menu?.clear()
         inflater?.inflate(R.menu.menu_share, menu)
-        if (FeatureFlag.PRODUCT_IMAGE_CHOOSER.isEnabled()) {
-            menu?.add(Menu.NONE, MENU_ID_CHOOSE_PHOTO, Menu.NONE, R.string.product_change_image)
-            menu?.add(Menu.NONE, MENU_ID_CAPTURE_PHOTO, Menu.NONE, R.string.product_capture_image)
-        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
@@ -181,14 +139,6 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
             R.id.menu_share -> {
                 AnalyticsTracker.track(PRODUCT_DETAIL_SHARE_BUTTON_TAPPED)
                 viewModel.onShareButtonClicked()
-                true
-            }
-            MENU_ID_CHOOSE_PHOTO -> {
-                viewModel.onChooseImageClicked()
-                true
-            }
-            MENU_ID_CAPTURE_PHOTO -> {
-                viewModel.onCaptureImageClicked()
                 true
             }
             else -> super.onOptionsItemSelected(item)
@@ -543,123 +493,9 @@ class ProductDetailFragment : BaseFragment(), OnGalleryImageClickListener {
         }
     }
 
-    private fun chooseProductImage() {
-        // only show the chooser if user already allowed storage permission, otherwise simply request the
-        // permission and do nothing else - this will be called again if the user then agrees to allow
-        // storage permission
-        if (requestStoragePermission()) {
-            val intent = Intent(Intent.ACTION_GET_CONTENT)
-            intent.type = "image/*"
-            val chooser = Intent.createChooser(intent, getString(R.string.product_change_image))
-            activity?.startActivityFromFragment(this, chooser, REQUEST_CODE_CHOOSE_PHOTO)
-        }
-    }
-
-    private fun captureProductImage() {
-        if (requestCameraPermission()) {
-            MediaUploadUtils.createCaptureImageIntent(requireActivity())?.let { intent ->
-                capturedPhotoUri = intent.getParcelableExtra(android.provider.MediaStore.EXTRA_OUTPUT)
-                requireActivity().startActivityFromFragment(this, intent, REQUEST_CODE_CAPTURE_PHOTO)
-            }
-        }
-    }
-
-    /**
-     * Triggered by the viewModel when an image is being uploaded or has finished uploading
-     */
-    private fun showUploadImageProgress(isUploading: Boolean) {
-        // TODO - for now we simply show a progress spinner in the middle of the gallery, a separate PR
-        // will tackle showing the progress on the image being replaced
-        imageUploadProgess.visibility = if (isUploading) View.VISIBLE else View.GONE
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == RESULT_OK) {
-            if (requestCode == REQUEST_CODE_CHOOSE_PHOTO && data != null) {
-                data.data?.let { imageUri ->
-                    viewModel.uploadProductMedia(navArgs.remoteProductId, imageUri)
-                }
-            } else if (requestCode == REQUEST_CODE_CAPTURE_PHOTO) {
-                capturedPhotoUri?.let { imageUri ->
-                    viewModel.uploadProductMedia(navArgs.remoteProductId, imageUri)
-                }
-            }
-        }
-    }
-
     override fun onGalleryImageClicked(image: WCProductImageModel, imageView: View) {
         AnalyticsTracker.track(PRODUCT_DETAIL_IMAGE_TAPPED)
         val action = ProductDetailFragmentDirections.actionProductDetailFragmentToProductImagesFragment(navArgs.remoteProductId)
         findNavController().navigate(action)
-    }
-
-    /**
-     * Requests storage permission, returns true only if permission is already available
-     */
-    private fun requestStoragePermission(): Boolean {
-        if (!isAdded) {
-            return false
-        } else if (WooPermissionUtils.hasStoragePermission(activity!!)) {
-            return true
-        }
-
-        val permissions = arrayOf(permission.WRITE_EXTERNAL_STORAGE)
-        requestPermissions(permissions, WooPermissionUtils.STORAGE_PERMISSION_REQUEST_CODE)
-        return false
-    }
-
-    /**
-     * Requests camera & storage permissions, returns true only if permissions are already
-     * available. Note that we need to ask for both permissions because we also need storage
-     * permission to store media from the camera.
-     */
-    private fun requestCameraPermission(): Boolean {
-        if (!isAdded) {
-            return false
-        }
-
-        val hasStorage = WooPermissionUtils.hasStoragePermission(activity!!)
-        val hasCamera = WooPermissionUtils.hasCameraPermission(activity!!)
-        if (hasStorage && hasCamera) {
-            return true
-        }
-
-        val permissions = when {
-            hasStorage -> arrayOf(permission.CAMERA)
-            hasCamera -> arrayOf(permission.WRITE_EXTERNAL_STORAGE)
-            else -> arrayOf(permission.CAMERA, permission.WRITE_EXTERNAL_STORAGE)
-        }
-
-        requestPermissions(
-                permissions,
-                WooPermissionUtils.CAMERA_PERMISSION_REQUEST_CODE
-        )
-        return false
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
-    ) {
-        if (!isAdded) {
-            return
-        }
-
-        val allGranted = WooPermissionUtils.setPermissionListAsked(
-                activity!!, requestCode, permissions, grantResults, checkForAlwaysDenied = true
-        )
-
-        if (allGranted) {
-            when (requestCode) {
-                WooPermissionUtils.STORAGE_PERMISSION_REQUEST_CODE -> {
-                    chooseProductImage()
-                }
-                WooPermissionUtils.CAMERA_PERMISSION_REQUEST_CODE -> {
-                    captureProductImage()
-                }
-            }
-        }
     }
 }
