+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductDetailViewModel.kt
@@ -1,15 +1,11 @@
 package com.woocommerce.android.ui.products
 
-import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import com.woocommerce.android.R
 import com.woocommerce.android.annotations.OpenClassOnDebug
 import com.woocommerce.android.di.UI_THREAD
-import com.woocommerce.android.media.MediaUploadService
-import com.woocommerce.android.media.MediaUploadService.Companion.OnProductMediaUploadEvent
-import com.woocommerce.android.media.MediaUploadWrapper
 import com.woocommerce.android.model.Product
 import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
@@ -18,9 +14,6 @@ import com.woocommerce.android.viewmodel.ScopedViewModel
 import com.woocommerce.android.viewmodel.SingleLiveEvent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
-import org.greenrobot.eventbus.EventBus
-import org.greenrobot.eventbus.Subscribe
-import org.greenrobot.eventbus.ThreadMode
 import org.wordpress.android.fluxc.store.WooCommerceStore
 import java.math.BigDecimal
 import javax.inject.Inject
@@ -34,8 +27,7 @@ class ProductDetailViewModel @Inject constructor(
     private val selectedSite: SelectedSite,
     private val productRepository: ProductDetailRepository,
     private val networkStatus: NetworkStatus,
-    private val currencyFormatter: CurrencyFormatter,
-    private val mediaUploadWrapper: MediaUploadWrapper
+    private val currencyFormatter: CurrencyFormatter
 ) : ScopedViewModel(mainDispatcher) {
     private var remoteProductId = 0L
 
@@ -51,17 +43,8 @@ class ProductDetailViewModel @Inject constructor(
     private val _shareProduct = SingleLiveEvent<Product>()
     val shareProduct: LiveData<Product> = _shareProduct
 
-    private val _chooseProductImage = SingleLiveEvent<Product>()
-    val chooseProductImage: LiveData<Product> = _chooseProductImage
-
-    private val _captureProductImage = SingleLiveEvent<Product>()
-    val captureProductImage: LiveData<Product> = _captureProductImage
-
     private val _showSnackbarMessage = SingleLiveEvent<Int>()
     val showSnackbarMessage: LiveData<Int> = _showSnackbarMessage
-
-    private val _isUploadingProductImage = MutableLiveData<Boolean>()
-    val isUploadingProductImage: LiveData<Boolean> = _isUploadingProductImage
 
     private val _exit = SingleLiveEvent<Unit>()
     val exit: LiveData<Unit> = _exit
@@ -77,8 +60,6 @@ class ProductDetailViewModel @Inject constructor(
                 _productData.value = combineData(prod, params)
             }
         }
-
-        EventBus.getDefault().register(this)
     }
 
     fun start(remoteProductId: Long) {
@@ -89,29 +70,14 @@ class ProductDetailViewModel @Inject constructor(
         _shareProduct.value = product.value
     }
 
-    fun onChooseImageClicked() {
-        _chooseProductImage.value = product.value
-    }
-
-    fun onCaptureImageClicked() {
-        _captureProductImage.value = product.value
-    }
-
     override fun onCleared() {
         super.onCleared()
 
         productRepository.onCleanup()
-        EventBus.getDefault().unregister(this)
-    }
-
-    fun reloadProduct() {
-        product.value = productRepository.getProduct(remoteProductId)
     }
 
     private fun loadProduct(remoteProductId: Long) {
         loadParameters()
-
-        _isUploadingProductImage.value = MediaUploadService.isUploadingForProduct(remoteProductId)
 
         val shouldFetch = remoteProductId != this.remoteProductId
         this.remoteProductId = remoteProductId
@@ -191,27 +157,6 @@ class ProductDetailViewModel @Inject constructor(
             number.toString()
         } else {
             int.toString()
-        }
-    }
-
-    fun uploadProductMedia(remoteProductId: Long, localImageUri: Uri) {
-        // TODO: at some point we want to support uploading multiple product images
-        if (MediaUploadService.isBusy()) {
-            _showSnackbarMessage.value = R.string.product_image_already_uploading
-            return
-        }
-        _isUploadingProductImage.value = true
-        mediaUploadWrapper.uploadProductMedia(remoteProductId, localImageUri)
-    }
-
-    @Suppress("unused")
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun onEventMainThread(event: OnProductMediaUploadEvent) {
-        _isUploadingProductImage.value = false
-        if (event.isError) {
-            _showSnackbarMessage.value = R.string.product_image_upload_error
-        } else {
-            reloadProduct()
         }
     }
