+++ WooCommerce/src/main/res/drawable/ic_gridicons_camera.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M17,12c0,1.7 -1.3,3 -3,3s-3,-1.3 -3,-3s1.3,-3 3,-3S17,10.3 17,12zM22,7v11c0,1.1 -0.9,2 -2,2H4c-1.1,0 -2,-0.9 -2,-2V7c0,-1.1 0.9,-2 2,-2V4h4v1h2l1,-2h6l1,2h2C21.1,5 22,5.9 22,7zM7.5,9c0,-0.8 -0.7,-1.5 -1.5,-1.5S4.5,8.2 4.5,9s0.7,1.5 1.5,1.5S7.5,9.8 7.5,9zM19,12c0,-2.8 -2.2,-5 -5,-5s-5,2.2 -5,5s2.2,5 5,5S19,14.8 19,12z"/>
+</vector>
