+++ WooCommerce/src/main/res/drawable/ic_gridicons_image.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M20,6L20,18C20,19.1046 19.1046,20 18,20L6,20C4.8954,20 4,19.1046 4,18L4,6C4,4.8954 4.8954,4 6,4L18,4C19.1046,4 20,4.8954 20,6ZM18,6L6,6L6,12.38L8.19,10.19L13.42,15.42L14.42,14.42C15.0495,13.9165 15.9562,13.9638 16.53,14.53L18,16L18,6ZM13,9.5C13,8.6716 13.6716,8 14.5,8C15.3284,8 16,8.6716 16,9.5C16,10.3284 15.3284,11 14.5,11C13.6716,11 13,10.3284 13,9.5Z"
+      android:fillColor="#000000"/>
+</vector>
