+++ WooCommerce/src/main/res/layout/fragment_product_images.xml
@@ -9,7 +9,7 @@
     tools:context="com.woocommerce.android.ui.products.ProductImagesFragment">
 
     <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/addPhotoButton"
+        android:id="@+id/addImageButton"
         style="@style/Woo.Button.Bordered"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -27,4 +27,13 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:isGridView="true" />
+
+    <ProgressBar
+        android:id="@+id/imageUploadProgess"
+        style="?android:attr/progressBarStyle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:visibility="gone"
+        tools:visibility="visible" />
 </LinearLayout>
